safari-534.53-branch/LayoutTests/fast/images/support-broken-image-delegate.html

<html>
<head>
<script>

if (window.layoutTestController) {
    layoutTestController.dumpResourceLoadCallbacks();
    layoutTestController.dumpAsText();
    layoutTestController.setShouldPaintBrokenImage(false);
    layoutTestController.waitUntilDone();
}

function loaded() {
    var result = document.getElementById('result');
    var img = document.getElementById('img');
    // The default broken image height in mac is 20.
    result.innerHTML = img.offsetHeight >= 10 ? "FAIL" : "PASS";
    if (window.layoutTestController) {
        layoutTestController.setShouldPaintBrokenImage(true);
        layoutTestController.notifyDone();
    }
}

function failed() {
    var result = document.getElementById('result');
    // Report 'FAIL' if WebCore didn't report onError on broken image.
    result.innerHTML = "FAIL";
    if (window.layoutTestController) {
        layoutTestController.setShouldPaintBrokenImage(true);
        layoutTestController.notifyDone();
    } 
}
</script>
<body>
Radar 8610908 - Add a resource delegate method to query if WebCore should paint the broken image.
<div id='img'>
<img id="broken" onError="loaded()" onLoad="failed()" src="resources/broken-image-with-invalid-format.png">
</div>
<div id='result'>
result text.
</div>
</body>
</html>
