LayoutTests/platform/chromium-win/fast/images/support-broken-image-delegate-expected.txt

CONSOLE MESSAGE: line 8: Uncaught TypeError: Object [object Object] has no method 'setShouldPaintBrokenImage'
resources/broken-image-with-invalid-format.png - willSendRequest <NSURLRequest URL resources/broken-image-with-invalid-format.png, main document URL support-broken-image-delegate.html, http method GET> redirectResponse (null)
resources/broken-image-with-invalid-format.png - didReceiveResponse <NSURLResponse resources/broken-image-with-invalid-format.png, http status code 0>
resources/broken-image-with-invalid-format.png - didFinishLoading
CONSOLE MESSAGE: line 18: Uncaught TypeError: Object [object Object] has no method 'setShouldPaintBrokenImage'
Radar 8610908 - Add a resource delegate method to query if WebCore should paint the broken image.

FAIL
